+++ packages/toolkit/config/webpack.config.js
@@ -1,8 +1,6 @@
 /**
  * Internal dependencies
  */
-const { resolve, join } = require('path');
-const DependencyExtractionWebpackPlugin = require('@wordpress/dependency-extraction-webpack-plugin');
 const {
 	getBuildFiles,
 	getTenUpScriptsConfig,
@@ -52,64 +50,18 @@ const config = {
 	defaultTargets,
 };
 
-const hasReactFastRefresh = projectConfig.hot && !isProduction;
-
-const sharedConfig = {
-	mode: 'development',
+module.exports = {
+	devtool: isProduction ? false : 'source-map',
+	mode,
+	devServer: getDevServer(config),
+	entry: getEntryPoints(config),
+	output: getOutput(config),
 	target: getTarget(config),
-	output: {
-		filename: '[name]/index.min.js',
-		path: resolve(process.cwd(), join('dist', 'fast-refresh')),
-	},
+	resolve: getResolve(config),
+	externals: getExternals(config),
+	performance: getPerfomance(config),
+	module: getModules(config),
+	plugins: getPlugins(config),
+	stats: getStats(config),
+	optimization: getOptimization(config),
 };
-
-const ReactRefreshConfig = hasReactFastRefresh
-	? [
-			{
-				...sharedConfig,
-				name: 'react-refresh-entry',
-				entry: {
-					'react-refresh-entry':
-						'@pmmmwh/react-refresh-webpack-plugin/client/ReactRefreshEntry.js',
-				},
-				plugins: [new DependencyExtractionWebpackPlugin()],
-			},
-			{
-				...sharedConfig,
-				name: 'react-refresh-runtime',
-				entry: {
-					'react-refresh-runtime': {
-						import: 'react-refresh/runtime.js',
-						library: {
-							name: 'ReactRefreshRuntime',
-							type: 'window',
-						},
-					},
-				},
-				plugins: [
-					new DependencyExtractionWebpackPlugin({
-						useDefaults: false,
-					}),
-				],
-			},
-	  ]
-	: [];
-
-module.exports = [
-	{
-		devtool: isProduction ? false : 'source-map',
-		mode,
-		devServer: getDevServer(config),
-		entry: getEntryPoints(config),
-		output: getOutput(config),
-		target: getTarget(config),
-		resolve: getResolve(config),
-		externals: getExternals(config),
-		performance: getPerfomance(config),
-		module: getModules(config),
-		plugins: getPlugins(config),
-		stats: getStats(config),
-		optimization: getOptimization(config),
-	},
-	...ReactRefreshConfig,
-];
