+++ packages/toolkit/scripts/start.js
@@ -25,25 +25,49 @@ if (hasWebpackConfig()) {
 	configPath = fromProjectRoot('webpack.config.js');
 }
 
-const config = require(configPath);
+const runWebpack = () => {
+	const config = require(configPath);
+	const compiler = webpack(config);
 
-const compiler = webpack(config);
+	const { devServer } = config;
 
-const { devServer } = config[0];
+	if (devServer) {
+		const devServerOptions = { ...devServer, open: false };
 
-if (devServer) {
-	const devServerOptions = { ...devServer, open: false };
+		const server = new WebpackDevServer(compiler, devServerOptions);
 
-	const server = new WebpackDevServer(compiler, devServerOptions);
+		server.listen(devServerOptions.port, '127.0.0.1');
+	} else {
+		compiler.watch(
+			{
+				aggregateTimeout: 600,
+			},
+			(err, stats) => {
+				displayWebpackStats(err, stats);
+			},
+		);
+	}
+};
 
-	server.listen(devServerOptions.port, '127.0.0.1');
+const hot = hasArgInCLI('--hot');
+
+if (hot) {
+	// compile the fast refresh bundle
+	const config = require(fromConfigRoot('webpack-fast-refresh.config.js'));
+	const compiler = webpack(config);
+	compiler.run((err, stats) => {
+		displayWebpackStats(err, stats);
+
+		compiler.close((closedErr) => {
+			if (closedErr) {
+				// eslint-disable-next-line no-console
+				console.error(closedErr);
+			} else {
+				// we can only call runWebpack after the compiler has closed
+				runWebpack();
+			}
+		});
+	});
 } else {
-	compiler.watch(
-		{
-			aggregateTimeout: 600,
-		},
-		(err, stats) => {
-			displayWebpackStats(err, stats);
-		},
-	);
+	runWebpack();
 }
