+++ packages/toolkit/utils/config.js
@@ -102,10 +102,11 @@ const hasEslintignoreConfig = () => hasProjectFile('.eslintignore');
 
 const getDefaultConfig = () => {
 	const wpMode = getArgFromCLI('--wp');
-	const devServer = hasArgInCLI('--dev-server');
+	const hot = hasArgInCLI('--hot');
+	// hot automatically enables dev server
+	const devServer = hasArgInCLI('--dev-server') || hot;
 	const devServerPort = Number(getArgFromCLI('--port')) || 8000;
 	const analyze = hasArgInCLI('--analyze');
-	const hot = hasArgInCLI('--hot');
 
 	return {
 		entry: require(fromConfigRoot('buildfiles.config.js')),
